Move new test into existing inmem eventStorageEngine test
Move new test into existing inmem eventStorageEngine test

#4824
